add StaticList

+++ Object.cpp
@@ -5,7 +5,7 @@
 namespace DSLib
 {
 
-void* Object::operator new (unsigned int size) throw()
+void* Object::operator new (size_t size) throw()
 {
     return malloc(size);
 }
@@ -15,7 +15,7 @@ void  Object::operator delete (void *p)
     free(p);
 }
    
-void * Object::operator new[] (unsigned int size) throw()
+void * Object::operator new[] (size_t size) throw()
 {
     return malloc(size)
 }
